requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
requests==2.32.3
pydantic==2.10.6
pyyaml==6.0.2
pydantic-settings==2.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.38
pytz==2025.1
tiktoken==0.8.0
openai==1.62.0
alembic==1.14.1
markitdown==0.0.1a5
aiofiles==24.1.0
playwright==1.50.0
#asyncpg==0.30.0
gunicorn==23.0.0
ruff==0.9.9
psycopg2-binary